Revert "Merge remote-tracking branch 'origin/Sprint-3-Demo' into LevelEditor"
This reverts commit 4cb683123389c6e66174858bb19ee2ee716a946e, reversing
changes made to 78c8fb5b1f6927810dce889a5337e469277fb8ab.

+++ Aenir Engine 3.0/src/AenirEngine/Graphics/Shader Files/CellShader/CellShaderPS.hlsl
@@ -8,8 +8,8 @@ cbuffer CB_CellShader {
 	float4 outerEdgeColor;
 	float4 glowColor;
 	float glowStr;
-	int innerEdgeThickness;
-	int outerEdgeThickness;
+	float innerEdgeThickness;
+	float outerEdgeThickness;
 	float specularPower;
 	float specularStrength;
 	float roughness;
